+++ gradle.properties
@@ -11,7 +11,7 @@ loader_version=0.15.10
 # Mod Properties
 mod_version=1.0.0
 maven_group=com.holebois
-archives_base_name=localmaps
+archives_base_name=client_maps
 
 # Dependencies
 fabric_version=0.97.0+1.20.4
