+++ src/main/java/com/holebois/client_maps/client_mapsMod.java
@@ -1,15 +1,15 @@
-package com.holebois.localmaps;
+package com.holebois.client_maps;
 
 import net.fabricmc.api.ModInitializer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class localmapsMod implements ModInitializer {
+public class client_mapsMod implements ModInitializer {
 	// This logger is used to write text to the console and the log file.
 	// It is considered best practice to use your mod id as the logger's name.
 	// That way, it's clear which mod wrote info, warnings, and errors.
-    public static final Logger LOGGER = LoggerFactory.getLogger("localmaps");
+    public static final Logger LOGGER = LoggerFactory.getLogger("client_maps");
 
 	@Override
 	public void onInitialize() {
