+++ src/main/java/com/holebois/client_maps/mixin/client_mapsMixin.java
@@ -1,4 +1,4 @@
-package com.holebois.localmaps.mixin;
+package com.holebois.client_maps.mixin;
 
 import net.minecraft.server.MinecraftServer;
 import org.spongepowered.asm.mixin.Mixin;
@@ -7,7 +7,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(MinecraftServer.class)
-public class localmapsMixin {
+public class client_mapsMixin {
 	@Inject(at = @At("HEAD"), method = "loadWorld")
 	private void init(CallbackInfo info) {
 		// This code is injected into the start of MinecraftServer.loadWorld()V
